requirements.txt
asgiref==3.9.1
attrs==25.3.0
Authlib==1.6.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
ecdsa==0.19.1
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Markdown==3.8.2
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
python-jose==3.5.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
